move ConstantGenerator to the Compositions module
