+++ models/blueprints/backwards/constant_generator.rb
@@ -1,0 +1,4 @@
+Syskit.warn_about_new_naming_convention
+module Rock
+    backward_compatible_constant :ConstantGenerator, 'Rock::Compositions::ConstantGenerator', 'rock/models/compositions/constant_generator'
+end
