+++ models/compositions.rb
@@ -1,0 +1,1 @@
+require "models/compositions/constant_generator.rb"
