--- test/suite_blueprints.rb
@@ -1,1 +1,0 @@
-require 'test/blueprints/test_constant_generator'
