+++ test/suite_compositions.rb
@@ -1,0 +1,1 @@
+require "test/compositions/test_constant_generator.rb"
